No assets analytics, wallet breakdown moved to smaller components
