+++ common/v2/features/Dashboard/components/WalletBreakdown/WalletBreakdownView.tsx
@@ -1,0 +1,219 @@
+import React from 'react';
+import styled from 'styled-components';
+
+import BreakdownChart from './BreakdownChart';
+import NoAssets from './NoAssets';
+import { Balance } from './types';
+
+import moreIcon from 'common/assets/images/icn-more.svg';
+
+const BreakDownHeading = styled.div`
+  margin: 0;
+  font-size: 20px;
+  font-weight: bold;
+  color: #424242;
+
+  @media (min-width: 1080px) {
+    font-size: 24px;
+  }
+`;
+
+const BreakDownHeadingExtra = styled.span`
+  display: inline;
+
+  @media (min-width: 1080px) {
+    display: none;
+  }
+`;
+
+const BreakDownChartWrapper = styled.div`
+  flex: 1;
+  padding-left: 15px;
+  padding-top: 15px;
+  padding-bottom: 15px;
+
+  @media (max-width: 1080px) {
+    padding-right: 15px;
+  }
+`;
+
+const PanelFigures = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-evenly;
+  margin: 30px 0;
+`;
+
+const PanelFigure = styled.div``;
+
+const PanelFigureValue = styled.div`
+  margin: 0;
+  font-size: 22px;
+  font-weight: bold;
+`;
+
+const PanelFigureLabel = styled.div`
+  margin: 0;
+  font-size: 16px;
+`;
+
+interface PanelDividerProps {
+  mobileOnly?: boolean;
+}
+
+const PanelDivider =
+  styled.div <
+  PanelDividerProps >
+  `
+  height: 1px;
+  margin-bottom: 15px;
+  background: #ddd;
+  display: block;
+
+  ${props =>
+    props.mobileOnly &&
+    `
+  @media (min-width: 1080px) {
+    display: none;
+  }`};
+`;
+
+const VerticalPanelDivider = styled.div`
+  width: 1px;
+  margin: 0 15px;
+  background: #ddd;
+  display: none;
+
+  @media (min-width: 1080px) {
+    display: block;
+  }
+`;
+
+const BreakDownBalances = styled.div`
+  flex: 1;
+  padding-right: 15px;
+  padding-top: 15px;
+  padding-bottom: 15px;
+
+  @media (max-width: 1080px) {
+    padding-left: 15px;
+  }
+`;
+
+const BreakDownHeadingWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+const BreakDownMore = styled.img`
+  display: block;
+  cursor: pointer;
+`;
+
+const BreakDownBalanceList = styled.div`
+  display: flex;
+  flex-direction: column;
+  color: #282d32;
+  font-size: 16px;
+  font-weight: normal;
+`;
+
+const BreakDownBalance = styled.div`
+  display: flex;
+  justify-content: space-between;
+  margin: 11px 0;
+  line-height: 1.2;
+
+  &:first-of-type {
+    margin-top: 16px;
+  }
+`;
+
+const BreakDownBalanceAssetName = styled.div`
+  margin: 0;
+`;
+
+const BreakDownBalanceAssetAmount = styled(BreakDownBalanceAssetName)`
+  a {
+    color: #1eb8e7;
+  }
+`;
+
+const BreakDownBalanceTotal = styled.div`
+  display: flex;
+  justify-content: space-between;
+  font-size: 16px;
+  font-weight: normal;
+`;
+
+interface WalletBreakdownProps {
+  balances: Balance[];
+  totalFiatValue: number;
+  toggleShowChart(): void;
+}
+
+export default function WalletBreakdownView({
+  balances,
+  toggleShowChart,
+  totalFiatValue
+}: WalletBreakdownProps) {
+  const highestPercentageAssetName = balances.length > 0 && balances[0].name;
+  const highestPercentage =
+    balances.length > 0 && Math.floor(balances[0].fiatValue / totalFiatValue * 100);
+
+  return balances.length === 0 ? (
+    <NoAssets />
+  ) : (
+    <>
+      <BreakDownChartWrapper>
+        <BreakDownHeading>
+          Wallet Breakdown <BreakDownHeadingExtra>(All Accounts)</BreakDownHeadingExtra>
+        </BreakDownHeading>
+        <BreakdownChart balances={balances} />
+        <PanelFigures>
+          <PanelFigure>
+            <PanelFigureValue>{highestPercentageAssetName}</PanelFigureValue>
+            <PanelFigureLabel>{highestPercentage}% Of Your Funds</PanelFigureLabel>
+          </PanelFigure>
+          <PanelFigure>
+            <PanelFigureValue>
+              ${balances.length > 0 && balances[0].fiatValue.toFixed(2)}
+            </PanelFigureValue>
+            <PanelFigureLabel>Value in US Dollars</PanelFigureLabel>
+          </PanelFigure>
+        </PanelFigures>
+      </BreakDownChartWrapper>
+      <PanelDivider mobileOnly={true} />
+      <VerticalPanelDivider />
+      <BreakDownBalances>
+        <BreakDownHeadingWrapper>
+          <BreakDownHeading>Balance</BreakDownHeading>
+          <BreakDownMore src={moreIcon} alt="More" onClick={toggleShowChart} />
+        </BreakDownHeadingWrapper>
+        <BreakDownBalanceList>
+          {balances.map(({ name, amount, fiatValue, ticker, isOther }) => (
+            <BreakDownBalance key={name}>
+              <div>
+                <BreakDownBalanceAssetName>{name}</BreakDownBalanceAssetName>
+                <BreakDownBalanceAssetAmount>
+                  {!isOther ? (
+                    `${amount.toFixed(4)} ${ticker}`
+                  ) : (
+                    <a onClick={toggleShowChart}>View Details</a>
+                  )}
+                </BreakDownBalanceAssetAmount>
+              </div>
+              <BreakDownBalanceAssetAmount>${fiatValue.toFixed(2)}</BreakDownBalanceAssetAmount>
+            </BreakDownBalance>
+          ))}
+          <PanelDivider />
+          <BreakDownBalanceTotal>
+            <div>Total</div>
+            <div>${totalFiatValue.toFixed(2)}</div>
+          </BreakDownBalanceTotal>
+        </BreakDownBalanceList>
+      </BreakDownBalances>
+    </>
+  );
+}
